test: point 동시성 테스트 속도 개선
- 병렬 처리 되도록 세 개의 파일로 분리
- 각 테스트 케이스 속도가 많이 느려서 동시 요청 개수 줄임 => 나중에 더 강력한 테스팅을 위해 다시 늘리는 것도 고려 가능
